+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Modifiers `firstVisible` and `lastVisible` for the `slide` CSS Handles 
+
 ## [0.11.0] - 2020-03-11
 ### Added
 - New CSS Handle `paginationDot--isActive`.